practice/src/main/java/ca/jrvs/practice/codingChallenge/MiddleOfLinkedList.java
package ca.jrvs.practice.codingChallenge;

/**
 * Ticket: https://www.notion.so/jarvisdev/Middle-of-the-Linked-List-b2cefd95507f4c05b6dfa7f5c2711513
 */
public class MiddleOfLinkedList {

  static class ListNode {

    int val;
    ListNode next;

    ListNode() {
    }

    ListNode(int val) {
      this.val = val;
    }

    ListNode(int val, ListNode next) {
      this.val = val;
      this.next = next;
    }

  }

  /**
   * Big-O: O(n)
   * Justification: While loop iterates at most n/2 times.
   */
  public ListNode middleNode(ListNode head) {
    ListNode slow = head, fast = head;
    while (fast != null && fast.next != null) {
      slow = slow.next;
      fast = fast.next.next;
    }
    return slow;
  }

}
